Created Splash page and removed UsersController
