+++ app/views/in/splash.html.php
@@ -1,0 +1,11 @@
+<div class="row">
+<div class="col-sm-4 col-sm-offset-1">
+<?php echo $this->_render('element', 'settings/two-factor'); ?>
+</div>
+</div>
+
+<div class="row col-sm-offset-5">
+<a href="/in/accounts/"><big>Skip >></big></a>
+</div>
+
+
